+++ src/main/java/kitchenpos/products/domain/JpaProductRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.products.domain;
 
+import kitchenpos.products.tobe.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
