+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -1,5 +1,7 @@
 package kitchenpos.products.domain;
 
+import kitchenpos.products.tobe.domain.Product;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
