+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.application;
 
-import kitchenpos.products.domain.Product;
+import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
 
 import java.util.*;
